refactor: rename the Mission file, add an Overview file
This is to be more clear about the intent of the files, since the new
mission statement will be much longer. The old mission page is more
closely related to an overview; thus, it is being renamed.
