+++ src/pages/index.ts
@@ -1,6 +1,7 @@
 export { default as Title } from './Title'
-export { default as Mission } from './Mission'
+export { default as Overview } from './Overview'
 export { default as Events } from './Events'
 export { default as SignUp } from './SignUp'
 export { default as Contact } from './Contact'
 export { default as Footer } from './Footer'
+export { default as Mission } from './Mission'
